refactor(process): migrate containercheck component to v2
