+++ comp/process/containercheck/def/component.go
